Fix build and add prefix parse in field
